graphExploration now ignores existing users correctly, added some documentation

+++ graphExploration.py
@@ -16,7 +16,7 @@ dummy = None
 session = requests.Session()
 q = queue.Queue()
 
-q.put(4611686018467544385)
+q.put(4611686018470746798) #Bamb
 
 edges = pd.DataFrame(columns=['from', 'to', 'weight'])
 vertices = pd.DataFrame(columns=['id', 'name'])
@@ -30,26 +30,35 @@ if not os.path.lexists('raiddata'):
 if not os.path.exists('raiddata/friendpickle'):
     os.mkdir('raiddata/friendpickle')
 
+
 while not q.empty():
     centerNodeDestinyID = q.get()
+    #loop while users already exist, find new one
     while os.path.exists(f'raiddata/friendpickle/{centerNodeDestinyID}.pickle'):
+        #add friends of existing member #BF
         for friend in pd.read_pickle(f"raiddata/friendpickle/{centerNodeDestinyID}.pickle")['to']:
-            q.put(friend)
-            #fill in lost names
-            if str(friend) not in vertices['id'].values:
-                for platform in [3,2,1,4,5,10,254]:
-                    charURL = "https://stats.bungie.net/Platform/Destiny2/{}/Profile/{}/?components=100,200"
-                    characterinfo = requests.get(url=charURL.format(platform,friend), headers=PARAMS).json()
-                    if not 'Response' in characterinfo:
-                        continue
-                    response = characterinfo['Response']
-                    name = response['profile']['data']['userInfo']['displayName']
-                    vertices = vertices.append({'id': str(friend),'name': name}, ignore_index= True)
-                    vertices.to_pickle("raiddata/vertices.pickle")
-                    print(f'added {name}:{friend} to vertices.pickle')
-                    break
+            #if friend isn't in the link-list yet
+            if not os.path.exists(f'raiddata/friendpickle/{friend}.pickle'):
+                #queue up for analysis
+                q.put(friend)
+            else:
+                #fill in lost names
+                if str(friend) not in vertices['id'].values:
+                    for platform in [3,2,1,4,5,10,254]:
+                        charURL = "https://stats.bungie.net/Platform/Destiny2/{}/Profile/{}/?components=100,200"
+                        characterinfo = requests.get(url=charURL.format(platform,friend), headers=PARAMS).json()
+                        if not 'Response' in characterinfo:
+                            continue
+                        response = characterinfo['Response']
+                        name = response['profile']['data']['userInfo']['displayName']
+                        vertices = vertices.append({'id': str(friend),'name': name}, ignore_index= True)
+                        vertices.to_pickle("raiddata/vertices.pickle")
+                        print(f'added {name}:{friend} to vertices.pickle')
+                        break
 
         centerNodeDestinyID = q.get()
+        #pull next from queue to analyse
+
     charURL = "https://stats.bungie.net/Platform/Destiny2/{}/Profile/{}/?components=100,200"
     characterinfo = None
     platform = 3
@@ -121,6 +130,14 @@ while not q.empty():
 
     #put them into a dataframe, group by unique users and count the common raids
     df = pd.DataFrame(partners, columns=['destinyid','username'])
+
+    #get people and names from df
+    vx = df.drop_duplicates(subset='destinyid')
+    vx_nicely_named = vx.rename(columns = {'destinyid':'id', 'username':'name'})
+    vertices = vertices.append(vx_nicely_named).drop_duplicates(subset='id')
+    vertices.to_pickle(f'raiddata/vertices.pickle')
+
+    #get edges and weights
     count = df.groupby(['destinyid','username']).size().reset_index(name='counts')
     sortedcount = count.sort_values('counts', ascending=False)
     sortedcount.drop(sortedcount.index[0], inplace=True)
@@ -139,11 +156,6 @@ while not q.empty():
     #edges = edges[edges['from'] > edges['to']]
     edges.to_pickle(f'raiddata/edges.pickle')
 
-
-    vx = df.drop_duplicates(subset='destinyid')
-    vx.columnns = ['id', 'name']
-    vertices = vertices.append(vx).drop_duplicates(subset='id')
-    vertices.to_pickle(f'raiddata/vertices.pickle')
 
     print(f'done with {centerNodeDestinyID}')
     for did in outputedge['to'].values:
